requirements.txt
darkdetect>=0.0.0,<1.0.0
psutil>=6.0.0,<7.0.0
py-cpuinfo>=9.0.0,<10.0.0
PySide6>=6.0.0,<7.0.0
PySide6-stubs>=6.0.0,<7.0.0
